Module 5: Accounts, Authentication and Permissions

+++ backend/src/resolvers/Mutation.js
@@ -1,12 +1,25 @@
 const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
+const { randomBytes } = require("crypto");
+const { promisify } = require("util");
+
+const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const Mutations = {
   createItem: async (parent, args, ctx, info) => {
-    // TODO: check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in to do that");
+    }
     const item = await ctx.db.mutation.createItem(
       {
         data: {
+          // This is how we create a relationship between item and user
+          user: {
+            connect: {
+              id: ctx.request.userId
+            }
+          },
           ...args
         }
       },
@@ -33,14 +46,19 @@ const Mutations = {
   deleteItem: async (parent, args, ctx, info) => {
     const where = { id: args.id };
     // 1.find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user{ id }}`);
     //2. check if they own that itme, or have the permissions
-    //TODO:
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+    if (!ownsItem && !hasPermission) {
+      throw new Error("You don't have the permission to do that!");
+    }
     //3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
   signup: async (parent, args, ctx, info) => {
-    console.log(args);
     args.email = args.email.toLowerCase();
     const password = await bcrypt.hash(args.password, 10);
     const user = await ctx.db.mutation.createUser(
@@ -72,9 +90,83 @@ const Mutations = {
     createAndSetJWTToken(ctx, user.id);
     return user;
   },
-  signout: async (parent, args, ctx, info) => {
+  signout: (parent, args, ctx, info) => {
     ctx.response.clearCookie("token");
     return { message: "Goodbye!" };
+  },
+  requestReset: async (parent, args, ctx, info) => {
+    //1. Check if this is a real user
+    const user = await ctx.db.query.user({ where: { email: args.email } });
+    if (!user) throw new Error(`No such user forund for email ${args.email}`);
+    // 2. Set a reset token and expiry on that user
+    const resetToken = (await promisify(randomBytes)(20)).toString("hex");
+    const resetTokenExpiry = Date.now + 360000; // 1 hour from now
+    const res = await ctx.db.mutation.updateUser({
+      where: { email: args.email },
+      data: { resetToken, resetTokenExpiry }
+    });
+    // TODO: 3. Email them that reset token
+    const mailRes = await transport.sendMail({
+      from: "carloslouredaparrado@gmail.com",
+      to: user.email,
+      subject: "Your password reset token",
+      html: makeANiceEmail(
+        `Your Password Reset Token is here! \n\n <a href="${
+          process.env.FRONTEND_URL
+        }/reset?resetToken=${resetToken}">Click here to Reset</a>`
+      )
+    });
+
+    return { message: "Thanks!" };
+  },
+  resetPassword: async (parent, args, ctx, info) => {
+    let { password, confirmPassword, resetToken } = args;
+    //1. Check if the passwords match
+    if (password !== confirmPassword) {
+      throw new Error("Oops! Passwords do not match.");
+    }
+    // 2 Check if its a legit reset token
+    // 3. Check i its expired
+    const [user] = await ctx.db.query.users({
+      where: { resetToken, resetTokenExpiry_gte: Date.now() - 360000 }
+    });
+    if (!user) throw new Error("Not valid reset password link or expired");
+    // 4. Hash their pasword
+    password = await bcrypt.hash(password, 10);
+    // 5. Save the new password to the user and remove old resettoekn
+    const updatedUser = await ctx.db.mutation.updateUser({
+      where: { email: user.email },
+      data: { password, resetToken: null, resetTokenExpiry: null }
+    });
+    // 6. Generate JWT
+    // 7. Set the JWT cookie
+    createAndSetJWTToken(ctx, updatedUser.id);
+    // 8. Return the new user
+    return updatedUser;
+  },
+  updatePermissions: async (parent, args, ctx, info) => {
+    console.log("[MUTATION] updatePermissions");
+    // 1. Check if they are logged in
+    if (!ctx.request.userId) throw new Error("You must be logged in!");
+    // 2. Query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+    // 3. Check if they have permissions to do this
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+    // 4. Update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: { permissions: { set: args.permissions } },
+        where: { id: args.userId }
+      },
+      info
+    );
   }
 };
